requirements.txt
ipykernel
langchain
python-dotenv
langgraph
youtube-transcript-api
streamlit
langchain_groq
pytube
gtts